fixed bugs in tictactoe/CORS
