+++ tictactoes/7-api-spa/api-services/board/controllers/boardController.js
@@ -2,14 +2,14 @@ var boardModel = require('../models/boardModel');
 
 var boardController = {
 	buildBoard: function(req, res) {
-		boardModel.buildBoard(w);
+		boardModel.buildBoard(req.body.data.boardDimension);
 		res.json({success: true});
 	},
 	addMove: function(req, res) {
 		var row = req.body.data.row;
 		var column = req.body.data.column;
 		var player = req.body.data.player;
-		console.log(boardModel.getBoardState())
+		//console.log(boardModel.getBoardState())
 		var moveStatus = boardModel.addMove(row, column, player);
 		if(moveStatus) {
 			res.json({success: false, error: 'pick an empty square'});
