+++ tictactoes/7-api-spa/package.json
@@ -18,7 +18,6 @@
     "vinyl-source-stream": "^1.1.0",
     "body-parser": "^1.17.2",
     "cors": "^2.8.4",
-    "express": "^4.15.3",
     "board": "file:api-services/board",
     "players": "file:api-services/players"
   }
