+++ tictactoes/7-api-spa/public/index.html~
@@ -3,7 +3,7 @@
 		<!-- prepare all things -->
 		
 		<!-- app script -->
-		<script src="./scripts/bundle.js"></script>
+		<script src="./scripts/main.js"></script>
 
 		<!-- axios, for async server calls -->
 		<script src="https://unpkg.com/axios/dist/axios.min.js"></script>
